+++ src/main/java/net/telepathicgrunt/bumblezone/configs/BzConfig.java
@@ -1,8 +1,8 @@
 package net.telepathicgrunt.bumblezone.configs;
 
-import me.sargunvohra.mcmods.autoconfig1u.ConfigData;
-import me.sargunvohra.mcmods.autoconfig1u.annotation.Config;
-import me.sargunvohra.mcmods.autoconfig1u.annotation.ConfigEntry;
+import me.shedaniel.autoconfig.ConfigData;
+import me.shedaniel.autoconfig.annotation.Config;
+import me.shedaniel.autoconfig.annotation.ConfigEntry;
 
 @Config(name = "the_bumblezone-main")
 public class BzConfig implements ConfigData {