+++ src/main/java/net/telepathicgrunt/bumblezone/configs/ModMenuConfig.java
@@ -3,7 +3,7 @@ package net.telepathicgrunt.bumblezone.configs;
 
 import io.github.prospector.modmenu.api.ConfigScreenFactory;
 import io.github.prospector.modmenu.api.ModMenuApi;
-import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
+import me.shedaniel.autoconfig.AutoConfig;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;